Edit some File

+++ pages/api/products/[id].js
@@ -5,7 +5,10 @@ export default async function handler(req, res) {
   const {
     method,
     query: { id },
+    cookies,
   } = req;
+
+  const token = cookies.token;
 
   dbConnect();
 
@@ -18,6 +21,9 @@ export default async function handler(req, res) {
     }
   }
   if (method === "PUT") {
+    if (!token || token !== process.env.TOKEN) {
+      return res.status(401).json("Not Authenticated!");
+    }
     try {
       const product = await Product.findByIdAndUpdate(req.body, id, {
         new: true,
@@ -29,6 +35,9 @@ export default async function handler(req, res) {
   }
 
   if (method === "DELETE") {
+    if (!token || token !== process.env.TOKEN) {
+      return res.status(401).json("Not Authenticated!");
+    }
     try {
       await Product.findByIdAndDelete(id);
       res.status(200).json("Product has been deleted!");
